WinGui: Fix ARM libhb detection

+++ win/CS/HandBrakeWPF/App.xaml.cs
@@ -59,9 +59,17 @@ namespace HandBrakeWPF
 
         private void Init(StartupEventArgs e)
         {
-            if (!File.Exists("hb.dll") && !File.Exists(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "hb.dll")))
+            if (!SystemInfo.IsArmDevice && !File.Exists("hb.dll") && !File.Exists(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "hb.dll")))
             {
                 MessageBox.Show("hb.dll file not found. Application will not run correctly without this. Please re-install HandBrake.", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                this.Shutdown();
+                Environment.Exit(-1);
+                return;
+            }
+
+            if (SystemInfo.IsArmDevice && !File.Exists("hb_a64.dll") && !File.Exists(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "hb_a64.dll")))
+            {
+                MessageBox.Show("hb_a64.dll file not found. Application will not run correctly without this. Please re-install HandBrake.", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
                 this.Shutdown();
                 Environment.Exit(-1);
                 return;
